Updated "update" command to use signal instead of flag file.
Previously, piPass monitored for an "update" request by reading from a flag
file on every iteration. This update removes that I/O activity and uses the
linux signal SIGUSR1 to catch update requests.

In addition to slightly reducing I/O load, this allows a change to the
STREETPASS_CYCLE_MINUTES setting to take effect for the current zone. If the
time is reduced below the current elapsed time, it will immediately advance.
Similarly if the time is extended, that extended time will apply to the current
zone.