Added check to skip recently visited zones.
When using the shuffle feature, it's very possible for zones near the end of
one pass to show up near the beginning of the next, potentially causing you to
hit a zone you visited too recently. This issue is now avoided by tracking the
time of the last visit and skipping a zone if it was seen too recently.

If there aren't enough zones to cycle through them all before the minimum 8
hour wait time, then the wait times will be cleared and visits will continue as
they had before.