PiPass 1.3 - RC3
NEW DEPENDENCY - PiPass relies on the installation of 7z (sudo apt-get
install p7zip-full).

PiPass can be upgraded through the Dashboard. It is a “rustic” upgrade
process in that it will always pull from the master branch and upgrade
PiPass to that version. PiPass settings are saved and restored as well.

Minor UI rearrangement and added comments to PHP files for maintenance
clarity.